Breakdown graphql country models
