+++ src/gql-api/models/countryNeighbour.model.ts
@@ -1,0 +1,11 @@
+import { Field, ObjectType } from '@nestjs/graphql';
+import { Country } from './country.model';
+
+@ObjectType({ description: 'Country neighbors' })
+export class CountryNeighours {
+  @Field(() => Country)
+  country: Country;
+
+  @Field(() => [Country])
+  neighbours: Country[];
+}
